<s>Gui cleanup. Minor testing implementations</s> My ability to hit commit never ceases to amaze me. GUI implemented, more or less. still working out a few bugs. Code converted to use callbacks instead of strings for buttons and output text, may revert that. base classes for general inventory system added. other changes. I did a lot and never hit commit. Oops.

+++ CoC.Frontend/Encounters/Common/EasterEggImpEncounter.cs
@@ -14,7 +14,7 @@ namespace CoC.Frontend.Encounters.Common
 {
 	class EasterEggImpEncounter : RandomEncounter
 	{
-		protected override int chances => Utils.LerpRound(1, 20, player.level, 2, 0);
+		protected override int chances => Utils.LerpRound(1, 20, (int)player.level, 2, 0);
 			
 		protected override bool encounterDisabled()
 		{
