+++ CoC.Frontend/Encounters/Common/GoblineAssassinEncounter.cs
@@ -14,7 +14,7 @@ namespace CoC.Frontend.Encounters.Common
 {
 	internal sealed class GoblinAssassinEncounter : RandomEncounter
 	{
-		protected override int chances => Utils.LerpRound(player.level, 10, player.level, 10, 40);
+		protected override int chances => Utils.LerpRound((int)player.level, 10, (int)player.level, 10, 40);
 
 		protected override bool encounterDisabled()
 		{
