+++ CoC.Frontend/Encounters/Common/ImpLordEncounter.cs
@@ -14,7 +14,7 @@ namespace CoC.Frontend.Encounters.Common
 {
 	internal sealed class ImpLordEncounter : RandomEncounter
 	{
-		protected override int chances => Utils.LerpRound(8, 16, player.level, 10, 40);
+		protected override int chances => Utils.LerpRound(8, 16, (int)player.level, 10, 40);
 
 		protected override bool encounterDisabled()
 		{
